request-scheduler/requirements.txt
flask_expects_json==1.5.0
Flask==1.1.2
requests==2.25.1
protobuf==3.14.0
python-decouple==3.4
google-cloud-pubsub==2.2.0
gunicorn==20.0.4